Make full use of the new charges table (#3993)
* Make full use of the new charges table

* minor enhancements
