+++ packages/client/src/components/charges-ledger-validation.tsx
@@ -2,12 +2,13 @@ import { useCallback, useContext, useEffect, useMemo, useState, type ReactElemen
 import { Check, Loader2, PanelTopClose, PanelTopOpen } from 'lucide-react';
 import { useQuery } from 'urql';
 import { Loader, Progress, ThemeIcon } from '@mantine/core';
+import type { RowSelectionState } from '@tanstack/react-table';
 import { encodeFilters, ROUTES } from '@/router/routes.js';
 import { ChargesLedgerValidationDocument, type ChargeFilter } from '../gql/graphql.js';
 import { useUrlQuery } from '../hooks/use-url-query.js';
 import { FiltersContext } from '../providers/filters-context.js';
 import { ChargesFilters } from './charges/charges-filters.js';
-import { ChargesTable } from './charges/charges-table.js';
+import { NewChargesTable } from './charges/new-charges-table.js';
 import { MergeChargesButton, Tooltip } from './common/index.js';
 import { PageLayout } from './layout/page-layout.js';
 import { Button } from './ui/button.js';
@@ -19,7 +20,7 @@ import { Button } from './ui/button.js';
       progress
       charge {
         id
-        ...ChargesTableFields
+        ...ChargeForChargesTableFields
       }
     }
   }
@@ -44,9 +45,7 @@ export function getLedgerValidationHref(filter?: ChargeFilter | null, page?: num
 export const ChargesLedgerValidation = (): ReactElement => {
   const { setFiltersContext } = useContext(FiltersContext);
   const [isAllOpened, setIsAllOpened] = useState<boolean>(false);
-  const [mergeSelectedCharges, setMergeSelectedCharges] = useState<
-    Array<{ id: string; onChange: () => void }>
-  >([]);
+  const [rowSelection, setRowSelection] = useState<RowSelectionState>({});
   const { get } = useUrlQuery();
   const uriFilters = get('chargesFilters');
   const initialFilters = useMemo(() => {
@@ -62,17 +61,6 @@ export const ChargesLedgerValidation = (): ReactElement => {
   }, [uriFilters]);
   const [filter, setFilter] = useState<ChargeFilter | undefined>(initialFilters);
 
-  const toggleMergeCharge = useCallback(
-    (chargeId: string, onChange: () => void) => {
-      if (mergeSelectedCharges.map(selected => selected.id).includes(chargeId)) {
-        setMergeSelectedCharges(mergeSelectedCharges.filter(selected => selected.id !== chargeId));
-      } else {
-        setMergeSelectedCharges([...mergeSelectedCharges, { id: chargeId, onChange }]);
-      }
-    },
-    [mergeSelectedCharges],
-  );
-
   const [{ data, fetching }, validateLedger] = useQuery({
     query: ChargesLedgerValidationDocument,
     variables: {
@@ -83,7 +71,7 @@ export const ChargesLedgerValidation = (): ReactElement => {
   });
 
   function onResetMerge(): void {
-    setMergeSelectedCharges([]);
+    setRowSelection({});
   }
 
   const progress = data?.chargesWithLedgerChanges?.length
@@ -102,6 +90,21 @@ export const ChargesLedgerValidation = (): ReactElement => {
       validateLedger();
     }
   }, [filter, validateLedger]);
+
+  // Derive the merge button's input from the row-selection map. Each selected charge gets an
+  // `onChange` that refetches the list, so the table refreshes once a merge completes.
+  const mergeSelectedCharges = useMemo(
+    () =>
+      Object.entries(rowSelection)
+        .filter(([, isSelected]) => isSelected)
+        .map(([id]) => ({
+          id,
+          onChange: (): void => {
+            validateLedger({ requestPolicy: 'network-only' });
+          },
+        })),
+    [rowSelection, validateLedger],
+  );
 
   useEffect(() => {
     setFiltersContext(
@@ -155,14 +158,14 @@ export const ChargesLedgerValidation = (): ReactElement => {
         <Loader2 className="h-10 w-10 animate-spin mr-2 self-center" />
       ) : (
         <>
-          <ChargesTable
-            toggleMergeCharge={toggleMergeCharge}
-            mergeSelectedCharges={new Set(mergeSelectedCharges.map(selected => selected.id))}
+          <NewChargesTable
+            rowSelection={rowSelection}
+            onRowSelectionChange={setRowSelection}
             data={
               data?.chargesWithLedgerChanges.filter(res => !!res.charge).map(res => res.charge!) ??
               []
             }
-            isAllOpened={isAllOpened}
+            // isAllOpened={isAllOpened}
           />
           <div className="flex flex-row justify-center my-2">
             {progress > 0 && progress < 100 && <Loader />}
