+++ packages/client/src/components/charges/cells/date.tsx
@@ -1,55 +1,55 @@
 import { type ReactElement } from 'react';
 import { format } from 'date-fns';
-import { ChargesTableDateFieldsFragmentDoc } from '../../../gql/graphql.js';
-import { getFragmentData, type FragmentType } from '../../../gql/index.js';
 
-// eslint-disable-next-line @typescript-eslint/no-unused-expressions -- used by codegen
-/* GraphQL */ `
-  fragment ChargesTableDateFields on Charge {
-    id
-    minEventDate
-    minDebitDate
-    minDocumentsDate
-    maxEventDate
-    maxDebitDate
-    maxDocumentsDate
+export function getDateProps({
+  minDebitDate,
+  minEventDate,
+  minDocumentsDate,
+  maxDebitDate,
+  maxEventDate,
+  maxDocumentsDate,
+}: {
+  minDebitDate: Date | null;
+  minEventDate: Date | null;
+  minDocumentsDate: Date | null;
+  maxDebitDate: Date | null;
+  maxEventDate: Date | null;
+  maxDocumentsDate: Date | null;
+}): DateProps | undefined {
+  if (!minDocumentsDate && !minEventDate && !minDebitDate) {
+    return undefined;
   }
-`;
-
-type Props = {
-  data: FragmentType<typeof ChargesTableDateFieldsFragmentDoc>;
-};
-
-export const DateCell = ({ data }: Props): ReactElement => {
-  const charge = getFragmentData(ChargesTableDateFieldsFragmentDoc, data);
-  const {
-    minDebitDate,
-    minEventDate,
-    minDocumentsDate,
-    maxDebitDate,
-    maxEventDate,
-    maxDocumentsDate,
-  } = charge;
-  const displayDate = minDocumentsDate || minEventDate || minDebitDate;
   const minTimestamps = [minDocumentsDate, minEventDate, minDebitDate]
     .filter(Boolean)
     .map(d => new Date(d!).getTime());
   const maxTimestamps = [maxDocumentsDate, maxEventDate, maxDebitDate]
     .filter(Boolean)
     .map(d => new Date(d!).getTime());
-  const mostMinDate = minTimestamps.length > 0 ? new Date(Math.min(...minTimestamps)) : null;
-  const mostMaxDate = maxTimestamps.length > 0 ? new Date(Math.max(...maxTimestamps)) : null;
+  const mostMinDate = minTimestamps.length > 0 ? new Date(Math.min(...minTimestamps)) : undefined;
+  const mostMaxDate = maxTimestamps.length > 0 ? new Date(Math.max(...maxTimestamps)) : undefined;
 
+  return {
+    date: minDocumentsDate || minEventDate || minDebitDate || undefined,
+    mostMinDate,
+    mostMaxDate,
+  };
+}
+
+export type DateProps = {
+  date?: Date;
+  mostMinDate?: Date;
+  mostMaxDate?: Date;
+};
+
+export const DateCell = ({ date, mostMinDate, mostMaxDate }: DateProps): ReactElement => {
   return (
-    <td>
-      <div>{displayDate && format(new Date(displayDate), 'dd/MM/yy')}</div>
-      <div className="text-xs text-gray-500">
-        {mostMinDate && mostMaxDate && mostMinDate.getTime() !== mostMaxDate.getTime() ? (
-          <>
-            ({format(mostMinDate, 'dd/MM/yy')} - {format(mostMaxDate, 'dd/MM/yy')})
-          </>
-        ) : null}
-      </div>
-    </td>
+    <>
+      <div>{date && format(date, 'dd/MM/yy')}</div>
+      {mostMinDate && mostMaxDate && mostMinDate.getTime() !== mostMaxDate.getTime() ? (
+        <div className="text-xs text-gray-500">
+          ({format(mostMinDate, 'dd/MM/yy')} - {format(mostMaxDate, 'dd/MM/yy')})
+        </div>
+      ) : null}
+    </>
   );
 };
