+++ packages/client/src/components/charges/cells/index.ts
@@ -1,4 +1,3 @@
-export { AccountantApproval } from './accountant-approval.js';
 export { Amount } from './amount.js';
 export { BusinessTrip } from './business-trip.js';
 export { Counterparty } from './counterparty.js';
