+++ packages/client/src/components/charges/cells/more-info.tsx
@@ -1,39 +1,31 @@
 import { useMemo, type ReactElement } from 'react';
 import { Indicator } from '@mantine/core';
-import { ChargesTableMoreInfoFieldsFragmentDoc, MissingChargeInfo } from '../../../gql/graphql.js';
-import { getFragmentData, type FragmentType } from '../../../gql/index.js';
+import type { ChargeType } from '../../../helpers/index.js';
 import { DragFile, ListCapsule } from '../../common/index.js';
 
-// eslint-disable-next-line @typescript-eslint/no-unused-expressions -- used by codegen
-/* GraphQL */ `
-  fragment ChargesTableMoreInfoFields on Charge {
-    __typename
-    id
-    metadata {
-      transactionsCount
-      documentsCount
-      ledgerCount
-      miscExpensesCount
-      ... on ChargeMetadata @defer {
-        invalidLedger
-      }
-    }
-    validationData {
-      missingInfo
-    }
-  }
-`;
-
-type Props = {
-  data: FragmentType<typeof ChargesTableMoreInfoFieldsFragmentDoc>;
+export type MoreInfoProps = {
+  chargeId: string;
+  type: ChargeType;
+  isTransactionsMissing?: boolean;
+  isDocumentsMissing?: boolean;
+  info?: {
+    transactionsCount: number;
+    documentsCount: number;
+    ledgerCount: number;
+    miscExpensesCount: number;
+    invalidLedger?: 'VALID' | 'DIFF' | 'INVALID';
+  };
 };
 
-export const MoreInfo = ({ data: rawData }: Props): ReactElement => {
-  const data = getFragmentData(ChargesTableMoreInfoFieldsFragmentDoc, rawData);
-  const { metadata, validationData, id, __typename } = data;
-
+export const MoreInfo = ({
+  chargeId,
+  type,
+  isTransactionsMissing,
+  isDocumentsMissing,
+  info,
+}: MoreInfoProps): ReactElement => {
   const shouldHaveDocuments = useMemo((): boolean => {
-    switch (__typename) {
+    switch (type) {
       case 'BusinessTripCharge':
       case 'ConversionCharge':
       case 'DividendCharge':
@@ -48,128 +40,101 @@ export const MoreInfo = ({ data: rawData }: Props): ReactElement => {
       default:
         return true;
     }
-  }, [__typename]);
+  }, [type]);
 
   const shouldHaveTransactions = useMemo((): boolean => {
-    switch (__typename) {
+    switch (type) {
       case 'FinancialCharge':
         return false;
       default:
         return true;
     }
-  }, [__typename]);
+  }, [type]);
 
   const isTransactionsError = useMemo(
-    () =>
-      shouldHaveTransactions &&
-      validationData?.missingInfo?.includes(MissingChargeInfo.Transactions),
-    [shouldHaveTransactions, validationData?.missingInfo],
+    () => shouldHaveTransactions && isTransactionsMissing,
+    [shouldHaveTransactions, isTransactionsMissing],
   );
 
   const isDocumentsError = useMemo(
-    () => shouldHaveDocuments && validationData?.missingInfo?.includes(MissingChargeInfo.Documents),
-    [shouldHaveDocuments, validationData?.missingInfo],
+    () => shouldHaveDocuments && isDocumentsMissing,
+    [shouldHaveDocuments, isDocumentsMissing],
   );
 
-  const isProcessingLedger = useMemo(() => !metadata?.invalidLedger, [metadata?.invalidLedger]);
-  const ledgerStatus = useMemo(() => metadata?.invalidLedger, [metadata?.invalidLedger]);
+  const ledgerStatus = useMemo(() => info?.invalidLedger, [info?.invalidLedger]);
 
-  const list = useMemo(() => {
-    const items: (
-      | React.ReactNode
-      | {
-          content: React.ReactNode;
-          extraClassName?: string;
-        }
-    )[] = [];
+  const list: (
+    | React.ReactNode
+    | {
+        content: React.ReactNode;
+        extraClassName?: string;
+      }
+  )[] = [];
 
-    if (isTransactionsError || metadata?.transactionsCount || shouldHaveTransactions) {
-      items.push({
-        extraClassName:
-          metadata?.transactionsCount || !shouldHaveTransactions ? undefined : 'bg-yellow-400',
-        content: (
-          <Indicator
-            key="transactions"
-            inline
-            size={12}
-            disabled={!isTransactionsError}
-            color="red"
-            zIndex="auto"
-          >
-            <div className="whitespace-nowrap">
-              Transactions: {metadata?.transactionsCount ?? 0}
-            </div>
-          </Indicator>
-        ),
-      });
-    }
-
-    items.push({
+  if (isTransactionsError || info?.transactionsCount || shouldHaveTransactions) {
+    list.push({
+      extraClassName:
+        info?.transactionsCount || !shouldHaveTransactions ? undefined : 'bg-yellow-400',
       content: (
         <Indicator
-          key="ledger"
+          key="transactions"
           inline
           size={12}
-          processing={isProcessingLedger}
-          disabled={ledgerStatus === 'VALID'}
-          color={ledgerStatus === 'DIFF' ? 'orange' : 'red'}
+          disabled={!isTransactionsError}
+          color="red"
           zIndex="auto"
         >
-          <div className="whitespace-nowrap">Ledger Records: {metadata?.ledgerCount ?? 0}</div>
+          <div className="whitespace-nowrap">Transactions: {info?.transactionsCount ?? 0}</div>
         </Indicator>
       ),
     });
+  }
 
-    if (isDocumentsError || metadata?.documentsCount) {
-      items.push({
-        content: (
-          <Indicator
-            key="documents"
-            inline
-            size={12}
-            disabled={!isDocumentsError}
-            color="red"
-            zIndex="auto"
-          >
-            <div className="whitespace-nowrap">Documents: {metadata?.documentsCount ?? 0}</div>
-          </Indicator>
-        ),
-        extraClassName:
-          !validationData?.missingInfo?.includes(MissingChargeInfo.Documents) ||
-          !shouldHaveDocuments
-            ? undefined
-            : 'bg-yellow-400',
-      });
-    }
+  list.push({
+    content: (
+      <Indicator
+        key="ledger"
+        inline
+        size={12}
+        processing={!ledgerStatus}
+        disabled={ledgerStatus === 'VALID'}
+        color={ledgerStatus === 'DIFF' ? 'orange' : 'red'}
+        zIndex="auto"
+      >
+        <div className="whitespace-nowrap">Ledger Records: {info?.ledgerCount ?? 0}</div>
+      </Indicator>
+    ),
+  });
 
-    if (metadata?.miscExpensesCount) {
-      items.push({
-        content: (
-          <div className="whitespace-nowrap">Misc Expenses: {metadata?.miscExpensesCount ?? 0}</div>
-        ),
-      });
-    }
+  if (isDocumentsError || info?.documentsCount) {
+    list.push({
+      content: (
+        <Indicator
+          key="documents"
+          inline
+          size={12}
+          disabled={!isDocumentsError}
+          color="red"
+          zIndex="auto"
+        >
+          <div className="whitespace-nowrap">Documents: {info?.documentsCount ?? 0}</div>
+        </Indicator>
+      ),
+      extraClassName: !isDocumentsMissing || !shouldHaveDocuments ? undefined : 'bg-yellow-400',
+    });
+  }
 
-    return items;
-  }, [
-    metadata?.transactionsCount,
-    metadata?.ledgerCount,
-    metadata?.documentsCount,
-    metadata?.miscExpensesCount,
-    shouldHaveDocuments,
-    shouldHaveTransactions,
-    isTransactionsError,
-    isDocumentsError,
-    validationData?.missingInfo,
-    ledgerStatus,
-    isProcessingLedger,
-  ]);
+  if (info?.miscExpensesCount) {
+    list.push({
+      content: (
+        <div className="whitespace-nowrap">Misc Expenses: {info.miscExpensesCount ?? 0}</div>
+      ),
+    });
+  }
 
   return (
-    <td>
-      <DragFile chargeId={id}>
-        <ListCapsule items={list} />
-      </DragFile>
-    </td>
+    <DragFile chargeId={chargeId}>
+      <ListCapsule items={list} />
+    </DragFile>
   );
 };
