+++ packages/client/src/components/charges/cells/tax-category.tsx
@@ -1,63 +1,18 @@
-import { useMemo, type ReactElement } from 'react';
+import { type ReactElement } from 'react';
 import { Indicator } from '@mantine/core';
-import {
-  ChargesTableTaxCategoryFieldsFragmentDoc,
-  MissingChargeInfo,
-} from '../../../gql/graphql.js';
-import { getFragmentData, type FragmentType } from '../../../gql/index.js';
 
-// eslint-disable-next-line @typescript-eslint/no-unused-expressions -- used by codegen
-/* GraphQL */ `
-  fragment ChargesTableTaxCategoryFields on Charge {
-    __typename
-    id
-    taxCategory {
-      id
-      name
-    }
-    validationData {
-      missingInfo
-    }
-  }
-`;
-
-type Props = {
-  data: FragmentType<typeof ChargesTableTaxCategoryFieldsFragmentDoc>;
+export type TaxCategoryProps = {
+  taxCategory?: {
+    id: string;
+    name: string;
+  };
+  isMissing?: boolean;
 };
 
-export const TaxCategory = ({ data }: Props): ReactElement => {
-  const { validationData, taxCategory, __typename } = getFragmentData(
-    ChargesTableTaxCategoryFieldsFragmentDoc,
-    data,
-  );
-
-  const shouldHaveTaxCategory = useMemo((): boolean => {
-    switch (__typename) {
-      case 'DividendCharge':
-      case 'InternalTransferCharge':
-      case 'SalaryCharge':
-      case 'BankDepositCharge':
-      case 'ForeignSecuritiesCharge':
-        return false;
-      default:
-        return true;
-    }
-  }, [__typename]);
-
-  const isError = useMemo(
-    () => validationData?.missingInfo?.includes(MissingChargeInfo.TaxCategory),
-    [validationData?.missingInfo],
-  );
-
-  if (!shouldHaveTaxCategory) {
-    return <td />;
-  }
-
+export const TaxCategory = ({ taxCategory, isMissing }: TaxCategoryProps): ReactElement => {
   return (
-    <td>
-      <Indicator inline size={12} disabled={!isError} color="red" zIndex="auto">
-        {taxCategory?.name ?? 'N/A'}
-      </Indicator>
-    </td>
+    <Indicator inline size={12} disabled={!isMissing} color="red" zIndex="auto">
+      {taxCategory?.name ?? 'N/A'}
+    </Indicator>
   );
 };
